Adapt trigger_service to OBS 2.5 behaviour.
It seems that packages in a linked project cannot be edited unless they are
first branched.
Detecting them is now a matter of determining that there is no <entry> inside the <directory> tag.
If such a package is found the link to the project being gated is extracted and the package is branched from there and then immediately overwritten using a new _service file.
It would be better to somehow just create a new not-linked package ready to take a _service file but it's not obvious how to do that as the package seems to exist already.

Signed-off-by: David Greaves <david.greaves@jolla.com>